move : jwtUtil클래스를 config -> util 폴더로 이동
