+++ src/main/java/com/example/workoutmate/global/config/JwtAuthenticationFilter.java
@@ -1,6 +1,7 @@
 package com.example.workoutmate.global.config;
 
 import com.example.workoutmate.domain.user.enums.UserRole;
+import com.example.workoutmate.global.util.JwtUtil;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.MalformedJwtException;
